feat: agent image info
